# Remove unnecessary methods from `setindex`