@ChrisRackauckas 2022-06-07
IIRC, this will break codes that need to be non-mutating on `Array` for compatibility with automatic differentiation. One example is trying to do Zygote over a code like https://github.com/JuliaDiff/FiniteDiff.jl/blob/cfee15e33394067f883fb685ea9f305d0bce9f0f/src/jacobians.jl#L196 . It at least was a missing function definition in Base and ideally these definitions would get upstreamed, maybe they already did. IIRC `setindex` used to just error by default on Arrays and it was a functionality that only Tuples and StaticArrays had?