@mateuszbaran 2022-06-07
Julia doesn't have this yet (see https://github.com/JuliaLang/julia/pull/33495 ). Anyway, that PR looks much better to me than the code here. Zygote compatibility should be easily possible by adding an rrule for `setindex`.