@ChrisRackauckas 2022-06-11
> Anyway, that PR looks much better to me than the code here.

It's not AD compatible so it would cause downstream failures.

> Zygote compatibility should be easily possible by adding an rrule for setindex.

That would be fine. If someone defines an rrule to delete these that's probably a better option. But I don't think we should merge something that we know is going to be downstream breaking like this, we should at least wait until a safer version is made, however it's done.