@hyrodium 2022-06-12
>That would be fine. If someone defines an rrule to delete these that's probably a better option. But I don't think we should merge something that we know is going to be downstream breaking like this, we should at least wait until a safer version is made, however it's done.

I'm not much familiar with Zygote.jl and related AD systems. I have some questions:

* Could you provide a MWE for the downstream breaking?
* Should we add some tests for AD and update ChainRulesTestUtils.jl in `[extras]` table in `lib/ArrayInterfaceCore/Project.toml`?
* Should we separate this PR into removing methods in `setindex` and AD-related things?