@hyrodium 2022-06-12
>One example is trying to do Zygote over a code like https://github.com/JuliaDiff/FiniteDiff.jl/blob/cfee15e33394067f883fb685ea9f305d0bce9f0f/src/jacobians.jl#L196 .

I have tested [FiniteDiff/test/runtests.jl](https://github.com/JuliaDiff/FiniteDiff.jl/blob/master/test/runtests.jl) with changes in this PR in my local PC, but the test didn't produce errors.