@ChrisRackauckas 2022-06-12
Rebase after https://github.com/JuliaArrays/ArrayInterface.jl/pull/313 for the new tests, and that might see the break. Otherwise I'll explicitly add something to show the break (but I'm traveling right now so it might now happen ASAP)