@ChrisRackauckas 2022-06-18
The simplest thing to do might be to add a ChainRules.rrule for `setindex` to ChainRules.jl. That should be rather simple to define, and if that merges then I'm 👍 on this. I still want to find the time to construct the case that hits this though, since it's a deep and nasty one to find/fix, but if we have the chain rule then it will go away.